Modify static variable and rename TextureManager.hpp
