+++ headers/Maze.hpp
@@ -21,7 +21,7 @@ public:
     void draw(sf::RenderWindow& window);
     bool move(Cell_mode mod, int old_x, int old_y, int new_x, int new_y);
     inline bool inside(int x, int y) const;
-    inline bool free_cell(int x, int y);// { return matrix[x][y].getMode() != WALL;}
+    bool free_cell(int x, int y);// { return matrix[x][y].getMode() != WALL;}
     friend std::ostream &operator<<(std::ostream &os, const Maze &maze);
 };
