+++ headers/Monsters/WanderingWhisk.hpp
@@ -19,8 +19,5 @@ public:
     void move(Maze &maze) override;
 };
 
-int WanderingWhisk::x_dir[4] = {0, 0, -1, 1};
-int WanderingWhisk::y_dir[4] = {-1, 1, 0, 0};
-Cell_mode WanderingWhisk::type = WANDERING_WHISK;
 
 #endif //OOP_WANDERINGWHISK_HPP
