+++ src/Monsters/BlinkingBanshee.cpp
@@ -4,6 +4,7 @@
 
 #include "../../headers/Monsters/BlinkingBanshee.hpp"
 
+Cell_mode BlinkingBanshee::type = BLINKING_BANSHEE;
 
 BlinkingBanshee::BlinkingBanshee(unsigned int x, unsigned int y) : Monster(x, y) {}
